removed unecesary trial files
