--- apps/auth-app/src/App.css
@@ -1,24 +1,0 @@
-.App {
-  text-align: center;
-  background-color: whitesmoke;
-}
-
-.App-header {
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-  justify-content: center;
-  font-size: calc(10px + 2vmin);
-  
-}
-.App-header2 {
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-  justify-content: center;
-  font-size: calc(8px + 2vmin);
-}
-.first{
-  align-items:center;
-  text-align: center;
-}
