--- apps/auth-app/tsconfig.json
@@ -1,29 +1,0 @@
-{
-  "extends": "../../tsconfig.common.json",
-  "compilerOptions": {
-    "lib": [
-      "dom",
-      "dom.iterable",
-      "esnext"
-    ],
-    "allowJs": true,
-    "allowSyntheticDefaultImports": true,
-    "skipLibCheck": true,
-    "esModuleInterop": true,
-    "strict": true,
-    "forceConsistentCasingInFileNames": true,
-    "moduleResolution": "node",
-    "resolveJsonModule": true,
-    "jsx": "react"
-  },
-  "include": [
-    "src"
-  ],
-  "exclude": [
-    "**/build/*",
-    "build",
-    "node_modules",
-    "**/*.test.*",
-    "webpack.config.js"
-  ],
-}
