--- apps/auth-app/webpack.config.js
@@ -1,51 +1,0 @@
-/* eslint-disable @typescript-eslint/no-var-requires */
-const path = require('path');
-const HtmlWebPackPlugin = require('html-webpack-plugin');
-
-module.exports = {
-  mode: 'production',
-  entry: './src/index.tsx',
-  module: {
-    rules: [
-      {
-        test: /\.jsx?$/,
-        exclude: /node_modules/,
-        use: {
-          loader: 'babel-loader',
-          options: {
-            presets: ['@babel/preset-env', '@babel/preset-react', '@babel/preset-typescript'],
-          },
-        },
-      },
-      {
-        test: /\.tsx?$/,
-        use: 'ts-loader',
-        exclude: /node_modules/,
-      },
-      {
-        test: /\.css$/,
-        use: ['style-loader', 'css-loader'],
-      },
-    ],
-  },
-  resolve: {
-    extensions: ['.tsx', '.ts', '.js', '.jsx'],
-  },
-  output: {
-    path: path.resolve(__dirname, 'build'),
-    filename: 'bundle.js',
-  },
-  devServer: {
-    contentBase: path.join(__dirname, 'build'),
-    publicPath: '/',
-    compress: true,
-    port: 4003,
-    https: true,
-    disableHostCheck: true,
-  },
-  optimization: {
-    minimize: false,
-  },
-  performance: { hints: false },
-  plugins: [new HtmlWebPackPlugin({ template: './index.html', filename: 'index.html' })],
-};
