+++ apps/sample-app/src/App.css
@@ -1,14 +1,24 @@
 .App {
   text-align: center;
+  background-color: whitesmoke;
 }
 
 .App-header {
-  background-color: #282c34;
-  min-height: 100vh;
   display: flex;
   flex-direction: column;
   align-items: center;
   justify-content: center;
   font-size: calc(10px + 2vmin);
-  color: white;
+  
+}
+.App-header2 {
+  display: flex;
+  flex-direction: column;
+  align-items: center;
+  justify-content: center;
+  font-size: calc(8px + 2vmin);
+}
+.first{
+  align-items:center;
+  text-align: center;
 }
