--- apps/sample-app/src/App0.tsx
@@ -1,15 +1,0 @@
-import './App.css';
-
-import React, { ReactElement } from 'react';
-
-const App0 = (): ReactElement => {
-  return (
-    <>
-      <div className="App">
-        <h1> Hello World </h1>
-      </div>
-    </>
-  );
-};
-
-export default App0;
