--- apps/sample-app/src/components/Dropdown.tsx
@@ -1,30 +1,0 @@
-import { Dropdown, DropdownMenuItemType, IDropdownOption, IDropdownStyles } from '@fluentui/react/lib/Dropdown';
-import { IStackTokens, Stack } from '@fluentui/react/lib/Stack';
-import * as React from 'react';
-
-const dropdownStyles: Partial<IDropdownStyles> = {
-  dropdown: { width: 300 },
-};
-
-const options: IDropdownOption[] = [
-  { key: 'fruitsHeader', text: 'Fruits', itemType: DropdownMenuItemType.Header },
-  { key: 'apple', text: 'Apple' },
-  { key: 'banana', text: 'Banana' },
-  { key: 'orange', text: 'Orange', disabled: true },
-  { key: 'grape', text: 'Grape' },
-  { key: 'divider_1', text: '-', itemType: DropdownMenuItemType.Divider },
-  { key: 'vegetablesHeader', text: 'Vegetables', itemType: DropdownMenuItemType.Header },
-  { key: 'broccoli', text: 'Broccoli' },
-  { key: 'carrot', text: 'Carrot' },
-  { key: 'lettuce', text: 'Lettuce' },
-];
-
-const stackTokens: IStackTokens = { childrenGap: 20 };
-
-export const DropdownBasicExample: React.FunctionComponent = () => {
-  return (
-    <Stack tokens={stackTokens}>
-      <Dropdown placeholder="Select an option" label="Trial Dropdown" options={options} styles={dropdownStyles} />
-    </Stack>
-  );
-};
