+++ apps/sample-app/src/components/EmailData.tsx
