+++ apps/sample-app/src/components/authConfig.js
